Fix: execucao NOx

+++ pipeline/2_anomalias.py
@@ -29,45 +29,38 @@ if __name__ == '__main__':
     abastecimento_exists = os.path.exists(path + "dados limpos/abastecimentos.csv")
 
     executeAbastecimento = os.getenv("EXECUTE_ABASTECIMENTO", "False").lower() == "true"
-    executeNox = os.getenv("EXECUTE_NOX","False").lower() == "true"    
+    executeNox = os.getenv("EXECUTE_NOX","False").lower() == "true"
 
-    if nox_exists and abastecimento_exists:
+    veiculos = pd.read_csv(path + "dados/informacoes_veiculos.csv", index_col=[0])
+
+    if executeNox and nox_exists:
         nox = pd.read_csv(path + "dados limpos/nox.csv", index_col=[0])
-        veiculos = pd.read_csv(path + "dados/informacoes_veiculos.csv", index_col=[0])
+        log_message("Found CSV files for NOx in 'dados_limpos'. Creating dataframes for NOx...")
+
+        clf = VoteEnsemble()
+        nox["anomaly_nox"] = clf.fit_predict(nox["NOx"])
+        nox["anomaly_o2"] = clf.fit_predict(nox["O2"])
+        nox = nox.replace({True: 1, False: 0})
+        if not os.path.exists(path + "anomalias/"): os.makedirs(path + "anomalias/")
+        nox.to_csv(path + "anomalias/nox.csv")
+        log_message('Printing DF nox:')
+        print(nox)
+
+    if executeAbastecimento and abastecimento_exists:
         abastecimento = pd.read_csv(path + "dados limpos/abastecimentos.csv", index_col=[0])
-        log_message("Found CSV files for NOx and Abastecimento in 'dados_limpos'. Creating dataframes")
+        log_message("Found CSV files for Abastecimento in 'dados_limpos'. Creating dataframes for Abastecimento...")
 
-        # Detecção Anomalias
-        log_message("Detecting Anomalies")
+        clf = VoteEnsemble()
+        abastecimento["anomaly_km_driven"] = clf.fit_predict(abastecimento["km_driven"])
+        abastecimento["anomaly_consumption"] = clf.fit_predict(abastecimento["consumption"])
+        abastecimento["anomaly_liter_supply"] = clf.fit_predict(abastecimento["liter_supply"])
+        abastecimento = abastecimento.replace({True: 1, False: 0})
         if not os.path.exists(path + "anomalias/"): os.makedirs(path + "anomalias/")
-        nox_exists = os.path.exists(path + "dados limpos/nox.csv")
-        abastecimento_exists = os.path.exists(path + "dados limpos/abastecimentos.csv")
+        abastecimento.to_csv(path + "anomalias/abastecimentos.csv")
+        log_message('Printing DF abastecimento:')
+        print(abastecimento)
 
-        # # Abastecimento
-        if executeAbastecimento:
-            clf = VoteEnsemble()
-            abastecimento["anomaly_km_driven"] = clf.fit_predict(abastecimento["km_driven"])
-            abastecimento["anomaly_consumption"] = clf.fit_predict(abastecimento["consumption"])
-            abastecimento["anomaly_liter_supply"] = clf.fit_predict(abastecimento["liter_supply"])
-            abastecimento = abastecimento.replace({True: 1, False: 0})
-            abastecimento.to_csv(path + "anomalias/abastecimentos.csv")
-            log_message('Printing DF abastecimento:')
-            print(abastecimento)
-
-        # # NOx
-        if not os.path.exists(path + "anomalias/"): os.makedirs(path + "anomalias/")
-        nox_exists = os.path.exists(path + "dados limpos/nox.csv")
-
-        if executeNox:
-            clf = VoteEnsemble()
-            nox["anomaly_nox"] = clf.fit_predict(nox["NOx"])
-            nox["anomaly_o2"] = clf.fit_predict(nox["O2"])
-            nox = nox.replace({True: 1, False: 0})
-            nox.to_csv(path + "anomalias/nox.csv")
-            log_message('Printing DF nox:')
-            print(nox)
-
-    else:
+    if (not abastecimento_exists) and (not nox_exists):
         log_error("Files for NOx and Abastecimento not found in 'dados_limpos'.")
